+++ packages/common/src/core/schemas/internal/getAuthedImageUrl.ts
@@ -1,17 +1,10 @@
 import { IArcGISContext } from "../../../ArcGISContext";
 import { cacheBustUrl } from "../../../urls/cacheBustUrl";
-import { EntityEditorOptions } from "./EditorOptions";
 
-export function getFeaturedImageUrl(
-  options: EntityEditorOptions,
-  context: IArcGISContext
-) {
+export function getAuthedImageUrl(url: string, context: IArcGISContext) {
   const queryParams = context.isAuthenticated
     ? `?token=${context.session.token}`
     : "";
   // TODO: Decide if the url should be passed in or plucked out of this deep path here
-  return (
-    options.view?.featuredImageUrl &&
-    cacheBustUrl(`${options.view.featuredImageUrl}${queryParams}`)
-  );
+  return url && cacheBustUrl(`${url}${queryParams}`);
 }