+++ packages/common/src/core/schemas/internal/getLocationExtent.ts
@@ -1,15 +1,16 @@
 import { bBoxToExtent, orgExtent } from "../../../extent";
 import { IHubRequestOptions } from "../../../types";
-import { EntityEditorOptions } from "./EditorOptions";
+import { IHubLocation } from "../../types";
+
 /**
  * Get the extent from the entity's location, if it has one.
  * Otherwise, fall back to using the org extent.
  */
 export async function getLocationExtent(
-  options: EntityEditorOptions,
+  location: IHubLocation,
   hubRequestOptions: IHubRequestOptions
 ) {
-  return options.location?.extent?.length
-    ? bBoxToExtent(options.location.extent)
+  return location?.extent?.length
+    ? bBoxToExtent(location.extent)
     : await orgExtent(hubRequestOptions);
 }