+++ packages/common/src/discussions/_internal/DiscussionUiSchemaEdit.ts
@@ -1,12 +1,12 @@
 import { IUiSchema } from "../../core/schemas/types";
 import { IArcGISContext } from "../../ArcGISContext";
-import { getFeaturedImageUrl } from "../../core/schemas/internal/getFeaturedImageUrl";
+import { getAuthedImageUrl } from "../../core/schemas/internal/getAuthedImageUrl";
 import { getTagItems } from "../../core/schemas/internal/getTagItems";
 import { getCategoryItems } from "../../core/schemas/internal/getCategoryItems";
 import { getLocationExtent } from "../../core/schemas/internal/getLocationExtent";
 import { getLocationOptions } from "../../core/schemas/internal/getLocationOptions";
 import { getThumbnailUiSchemaElement } from "../../core/schemas/internal/getThumbnailUiSchemaElement";
-import { EntityEditorOptions } from "../../core/schemas/internal/EditorOptions";
+import { IHubDiscussion } from "../../core/types";
 
 /**
  * @private
@@ -16,7 +16,7 @@ import { EntityEditorOptions } from "../../core/schemas/internal/EditorOptions";
  */
 export const buildUiSchema = async (
   i18nScope: string,
-  options: EntityEditorOptions,
+  options: Partial<IHubDiscussion>,
   context: IArcGISContext
 ): Promise<IUiSchema> => {
   return {
@@ -98,14 +98,21 @@ export const buildUiSchema = async (
               },
             },
           },
-          getThumbnailUiSchemaElement(i18nScope, options),
+          getThumbnailUiSchemaElement(
+            i18nScope,
+            options.thumbnail,
+            options.thumbnailUrl
+          ),
           {
             labelKey: `${i18nScope}.fields.featuredImage.label`,
             scope: "/properties/view/properties/featuredImage",
             type: "Control",
             options: {
               control: "hub-field-input-image-picker",
-              imgSrc: getFeaturedImageUrl(options, context),
+              imgSrc: getAuthedImageUrl(
+                options.view?.featuredImageUrl,
+                context
+              ),
               maxWidth: 727,
               maxHeight: 484,
               aspectRatio: 1.5,
@@ -136,7 +143,7 @@ export const buildUiSchema = async (
             options: {
               control: "hub-field-input-combobox",
               items: await getTagItems(
-                options,
+                options.tags,
                 context.portal.id,
                 context.hubRequestOptions
               ),
@@ -177,11 +184,13 @@ export const buildUiSchema = async (
             options: {
               control: "hub-field-input-location-picker",
               extent: await getLocationExtent(
-                options,
+                options.location,
                 context.hubRequestOptions
               ),
               options: await getLocationOptions(
-                options,
+                options.id,
+                options.type,
+                options.location,
                 context.portal.name,
                 context.hubRequestOptions
               ),