+++ packages/common/src/templates/_internal/TemplateUiSchemaEdit.ts
@@ -103,7 +103,11 @@ export const buildUiSchema = async (
               },
             },
           },
-          getThumbnailUiSchemaElement(i18nScope, options),
+          getThumbnailUiSchemaElement(
+            i18nScope,
+            options.thumbnail,
+            options.thumbnailUrl
+          ),
         ],
       },
     ],